Handle out of memory by max

+++ library/Image.class.php
@@ -70,11 +70,11 @@ class Image {
 		    
 		    // On vérifie l'extension
 		    if(substr($url, -3) == 'jpg' OR substr($url, -4) == 'jpeg') {
-			    // Téléchargement de l'image dans le dossier temp
-			    copy($url, FOLDER_IMAGES.'/temp/'.$ficheid.'.jpg');
+		    // Téléchargement de l'image dans le dossier temp
+		    copy($url, FOLDER_IMAGES.'/temp/'.$ficheid.'.jpg');
 
-			    // On redimensionne l'image pour qu'elle respecte les bonnes proportions
-			    $this->imageResize(FOLDER_IMAGES.'/temp/'.$ficheid.'.jpg', FOLDER_IMAGES.'/'.$fichetype.'/'.$this->folder($ficheid), $ficheid, 350, 480);
+		    // On redimensionne l'image pour qu'elle respecte les bonnes proportions
+		    $this->imageResize(FOLDER_IMAGES.'/temp/'.$ficheid.'.jpg', FOLDER_IMAGES.'/'.$fichetype.'/'.$this->folder($ficheid), $ficheid, 350, 480);
 			}
 		}
 	}
@@ -90,24 +90,30 @@ class Image {
 	  * @return void
 	  */
 	 public function imageResize($url, $folder, $nom, $largeur, $hauteur) {
-		$source = imagecreatefromjpeg($url); // La photo est la source
-		$largeur_source = imagesx($source);
-		$hauteur_source = imagesy($source);
+		if(filesize($url) < 1100000) {
+			echo '<p>'.$nom.' : OK '.filesize($url).'</p>';
+			$source = imagecreatefromjpeg($url); // La photo est la source
+			$largeur_source = imagesx($source);
+			$hauteur_source = imagesy($source);
 
-		$image = imagecreatetruecolor($largeur, $hauteur);
-		imagecopyresampled($image, $source, 0, 0, 0, 0, $largeur, $hauteur, $largeur_source, $hauteur_source);
-		
+			$image = imagecreatetruecolor($largeur, $hauteur);
+			imagecopyresampled($image, $source, 0, 0, 0, 0, $largeur, $hauteur, $largeur_source, $hauteur_source);
+			
+			imagedestroy($source);
 
-		// On supprime l'image de destination si elle existe
-		if(file_exists($folder.'/'.$nom.'.jpg'))
-		  unlink($folder.'/'.$nom.'.jpg');
+			// On supprime l'image de destination si elle existe
+			if(file_exists($folder.'/'.$nom.'.jpg'))
+				unlink($folder.'/'.$nom.'.jpg');
 
-		// On enregistre l'image redimensionnée
-		imagejpeg($image, $folder.'/'.$nom.'.jpg', 100);
+			// On enregistre l'image redimensionnée
+			imagejpeg($image, $folder.'/'.$nom.'.jpg', 100);
 
-		// On supprime l'image source
-		if(file_exists($url))
-		  unlink($url);
+			// On supprime l'image source
+			if(file_exists($url))
+				unlink($url);
+
+			imagedestroy($image);
+		}
 	}
 
 	public static function getUrl($type, $id) {
